fix(overlay): give the terminal a selection colour it can read
Same root cause as ba0c8bdf, one surface over: the light theme writes
--accent-dim as color-mix(), and xterm's parser throws on any colour whose
canvas read-back is translucent, so ThemeService quietly swapped in its own
default and the selection lost the accent tint it was supposed to have.

The resolver moves out of the interactive-artifact module and into the theme
service, which is the layer that owns turning a palette token into something
paintable; both the candlestick chart and the terminal now read it from there.
artifactVisualTheme drops its inlined copy of the raw token reader while it is
in the neighbourhood.
